requirements.txt
fastapi
uvicorn
openai
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
pyyaml
gradio
